cli: Refactor logics of commands into one
